Finsihed with hero section of home screen
