+++ src/components/Projects/projects-info.tsx
@@ -1,0 +1,18 @@
+export const ProjectItems = () => {
+  return [
+    { 
+      title: "2024 GMTK GameJam",
+      year: "2024",
+      description: "A turn base, dice rolling, dungeon crawler where you assemble an arsenal of spells to defeat oversized game pieces",
+      results: [
+        { bullet: "" },
+        { bullet: "Build under 4 days" },
+      ],
+      button: [
+        { name: "GitHub", link: "" },
+        { name: "Link", link: "" },
+      ],
+      image: "",
+    },
+  ];
+};
